Have CoreCLR follow sequential layout with GC refs
By default CoreCLR will ignore the LayoutKind.Sequential on
types containing GC references since those types are not
blittable and no code outside of the runtime can access those types
and will use the auto layout algorithm to change the field layout.

But Unity uses a lot of custom icalls that rely on the managed layout
matching the native layout.

This PR adds an IsSequentialWithRefs flag on MethodTable to indicate
the types with sequential layout, but we GC refs.  These types:

1. Have the LayoutKind.Sequential flag
2. The type is not blittable and not managed sequential (not blittable but has no GC refs).  These are already handled specially by the runtime.
3. Has a parent that is System.Object, System.ValueType, is sequential with refs, is blittable, or is managed sequential

This requires only very small changes in the runtime because the explicit layout paths do the work we need already.

1. classlayoutinfo.cpp already calculates field size and sequential offsets so we only need to mark the types as sequential with refs there.
2. In methodtablebuilder.cpp we need to set the field offset to the value calculated in 1.
3. In methodtablebuilder.cpp we run HandleExplicitLayout() which will calculate the type's size (it also, needlessly in our case, validates the field offsets)
4. In methodtablebuilder.cpp we run HandleGCForExplicitLayout() which calculates the GCDesc for the type.

The only other needed change was to change the definition of MetadataEnumResult in managedmdimport.hpp.
This type is used in an icall that assumes the auto-layout algorithm - interestingly it has nothing to do with GC
refs, but the fact that the auto-layout algorithm places larger than pointer sized structs on pointer sized boundaries.

This PR also updates CrossGen2 so follows the same layout algorithm.  The runtime has a lot of checks that
validate that the R2R image match the current runtime layout, size, and GC desc.